requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
dj-database-url==2.1.0
Django==5.0.4
django-bootstrap5==24.1
django-extensions==3.2.3
django-user-agents==0.4.0
djangorestframework==3.15.1
Faker==24.9.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
Markdown==3.6
pillow==10.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reportlab==4.1.0
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
websockets==12.0
whitenoise==6.6.0
zope.interface==6.3
